Safari-611.1.10/LayoutTests/canvas/philip/tests/2d.text.font.parse.size.percentage.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.font.parse.size.percentage</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" style="font-size: 144px" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.font = '50% serif';
_assertSame(ctx.font, '72px serif', "ctx.font", "'72px serif'");
canvas.setAttribute('style', 'font-size: 100px');
_assertSame(ctx.font, '72px serif', "ctx.font", "'72px serif'");


});
</script>
